+++ Makefile
@@ -3,4 +3,6 @@ test:
 	python -m unittest -v tests/test_*.py
 
 publish:
-	echo publishing to pypi
+	pip install python-semantic-release
+	semantic-release publish
+
