+++ pluralizer/__init__.py
@@ -1,0 +1,4 @@
+__version__ = "0.1.0"
+__license__ = "MIT"
+
+from .pluralizer import Pluralizer
