+++ pluralizer/pluralizer_rules.py
