+++ setup.cfg
@@ -1,0 +1,2 @@
+[semantic_release]
+version_variable=pluralizer/__init__.py:__version__
